requirements.txt
discord.py>=2.0.0
flask>=2.0.0
apscheduler>=3.9.0
python-dateutil>=2.8.0
werkzeug>=2.0.0
python-dotenv>=1.0.0
flask-wtf>=1.1.0
flask-limiter>=3.0.0